pyproject.toml
[project]
name = "agentic-converter"
version = "1.0.0"
description = "CLI tool that converts Jenkinsfiles to GitHub Actions YAML using a local LLM and an iterative agentic loop"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
